restructured our code, added feature for filtering top rated restaurants(Episode 5)
