+++ namastereact/foodorderapp/src/App.css
@@ -17,10 +17,13 @@
   border: 1px solid black;
 }
 .card-container {
-  border: 1px solid black;
   width: 300px;
   margin: 10px;
   padding: 5px;
+  background-color: #d3d3d3;
+}
+.container :hover {
+  border: 1px solid black;
 }
 .card-img {
   width: 100%;
@@ -29,4 +32,7 @@
 .card {
   display: flex;
   flex-wrap: wrap;
+}
+.container-btn {
+  margin: 10px;
 }
