+++ namastereact/foodorderapp/src/components/utils/mockData.js
@@ -1,0 +1,1256 @@
+const dataList = [
+  {
+    info: {
+      id: "21001",
+      name: "Pizza Hut",
+      cloudinaryImageId: "2b4f62d606d1b2bfba9ba9e5386fabb7",
+      locality: "F.C. Road",
+      areaName: "Shivajinagar",
+      costForTwo: "₹350 for two",
+      cuisines: ["Pizzas"],
+      avgRating: 4,
+      parentId: "721",
+      avgRatingString: "4.0",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 41,
+        lastMileTravel: 2,
+        serviceability: "SERVICEABLE",
+        slaString: "41 mins",
+        lastMileTravelString: "2.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-18 02:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/pizza-hut-f-c-road-shivajinagar-pune-21001",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "394784",
+      name: "Theobroma",
+      cloudinaryImageId: "b033728dcb0101ceb19bff0e1e1f6474",
+      locality: "Shivaji Nagar",
+      areaName: "Shivajinagar",
+      costForTwo: "₹400 for two",
+      cuisines: ["Bakery", "Desserts"],
+      avgRating: 4.6,
+      parentId: "1040",
+      avgRatingString: "4.6",
+      totalRatingsString: "1K+",
+      sla: {
+        deliveryTime: 36,
+        lastMileTravel: 2.8,
+        serviceability: "SERVICEABLE",
+        slaString: "36 mins",
+        lastMileTravelString: "2.8 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:59:00",
+        opened: true,
+      },
+      badges: {
+        imageBadges: [
+          {
+            imageId: "newg.png",
+            description: "Gourmet",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "Gourmet",
+                  imageId: "newg.png",
+                },
+              },
+            ],
+          },
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/theobroma-shivaji-nagar-shivajinagar-pune-394784",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "42581",
+      name: "Burger King",
+      cloudinaryImageId: "e33e1d3ba7d6b2bb0d45e1001b731fcf",
+      locality: "The Pavillion Mall",
+      areaName: "Shivajinagar",
+      costForTwo: "₹350 for two",
+      cuisines: ["Burgers", "American"],
+      avgRating: 4.1,
+      parentId: "166",
+      avgRatingString: "4.1",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 50,
+        lastMileTravel: 4.8,
+        serviceability: "SERVICEABLE",
+        slaString: "50 mins",
+        lastMileTravelString: "4.8 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-18 02:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/burger-king-the-pavillion-mall-shivajinagar-pune-42581",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "288388",
+      name: "Starbucks Coffee",
+      cloudinaryImageId: "258fe8a3577877fbfe064095ed1d9dc3",
+      locality: "FC Road",
+      areaName: "Shivajinagar",
+      costForTwo: "₹400 for two",
+      cuisines: ["Beverages", "Cafe", "Snacks", "Desserts", "Bakery", "Ice Cream"],
+      avgRating: 4.4,
+      parentId: "195515",
+      avgRatingString: "4.4",
+      totalRatingsString: "1K+",
+      sla: {
+        deliveryTime: 33,
+        lastMileTravel: 2.2,
+        serviceability: "SERVICEABLE",
+        slaString: "33 mins",
+        lastMileTravelString: "2.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:00:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/starbucks-coffee-fc-road-shivajinagar-pune-288388",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "23719",
+      name: "McDonald's",
+      cloudinaryImageId: "405645b3118d83e89db4c65361e43733",
+      locality: "JM Road",
+      areaName: "Shivajinagar",
+      costForTwo: "₹400 for two",
+      cuisines: ["Burgers", "Beverages", "Cafe", "Desserts"],
+      avgRating: 4.5,
+      parentId: "630",
+      avgRatingString: "4.5",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 32,
+        lastMileTravel: 2.1,
+        serviceability: "SERVICEABLE",
+        slaString: "32 mins",
+        lastMileTravelString: "2.1 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-18 01:45:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/mcdonalds-jm-road-shivajinagar-pune-23719",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "243625",
+      name: "KFC",
+      cloudinaryImageId: "f01666ac73626461d7455d9c24005cd4",
+      locality: "Deccan Gymkhana",
+      areaName: "Deccan Gymkhana",
+      costForTwo: "₹400 for two",
+      cuisines: ["Burgers", "Biryani", "American", "Snacks", "Fast Food"],
+      avgRating: 4.1,
+      parentId: "547",
+      avgRatingString: "4.1",
+      totalRatingsString: "1K+",
+      sla: {
+        deliveryTime: 33,
+        lastMileTravel: 2.2,
+        serviceability: "SERVICEABLE",
+        slaString: "33 mins",
+        lastMileTravelString: "2.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-18 02:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/kfc-deccan-gymkhana-pune-243625",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "5370",
+      name: "Cafe Goodluck",
+      cloudinaryImageId: "jsxy3zkxvxzkuqqpvjgd",
+      locality: "Deccan Gymkhana",
+      areaName: "Deccan Gymkhana",
+      costForTwo: "₹350 for two",
+      cuisines: ["North Indian", "Biryani", "Snacks", "Beverages", "Bakery"],
+      avgRating: 4.5,
+      parentId: "6539",
+      avgRatingString: "4.5",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 32,
+        lastMileTravel: 2.2,
+        serviceability: "SERVICEABLE",
+        slaString: "32 mins",
+        lastMileTravelString: "2.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/cafe-goodluck-deccan-gymkhana-pune-5370",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "29639",
+      name: "Eatsome- Wraps & Rolls",
+      cloudinaryImageId: "e56240a4b58956f47a5a1f8392470fbe",
+      locality: "J.M.Road",
+      areaName: "Shivajinagar",
+      costForTwo: "₹300 for two",
+      cuisines: ["North Indian", "Biryani", "Street Food", "Beverages"],
+      avgRating: 4.2,
+      parentId: "471587",
+      avgRatingString: "4.2",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 30,
+        lastMileTravel: 1.7,
+        serviceability: "SERVICEABLE",
+        slaString: "30 mins",
+        lastMileTravelString: "1.7 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/eatsome-wraps-and-rolls-j-m-road-shivajinagar-pune-29639",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "5743",
+      name: "Madinah",
+      cloudinaryImageId: "n07bhbglbqthyd45clul",
+      locality: "Camp",
+      areaName: "Camp",
+      costForTwo: "₹350 for two",
+      cuisines: ["Mughlai", "North Indian", "Chinese", "Biryani"],
+      avgRating: 4,
+      parentId: "1397",
+      avgRatingString: "4.0",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 36,
+        lastMileTravel: 3,
+        serviceability: "SERVICEABLE",
+        slaString: "36 mins",
+        lastMileTravelString: "3.0 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:30:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/madinah-camp-pune-5743",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "13909",
+      name: "Wadeshwar",
+      cloudinaryImageId: "uwny9zuejmpvoznipn3c",
+      locality: "Deccan Gymkhana",
+      areaName: "Koramangala",
+      costForTwo: "₹350 for two",
+      cuisines: ["South Indian", "Street Food", "Snacks", "Beverages", "Healthy Food", "Indian"],
+      avgRating: 4.6,
+      parentId: "1770",
+      avgRatingString: "4.6",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 37,
+        lastMileTravel: 2.5,
+        serviceability: "SERVICEABLE",
+        slaString: "37 mins",
+        lastMileTravelString: "2.5 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/wadeshwar-deccan-gymkhana-koramangala-pune-13909",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "35370",
+      name: "Falahaar",
+      cloudinaryImageId: "ynt91kk0vyornubvrdgl",
+      locality: "The pavillion mall",
+      areaName: "Shivajinagar",
+      costForTwo: "₹350 for two",
+      cuisines: ["Beverages", "Juices", "Fast Food", "Snacks", "Pizzas", "Italian", "Street Food", "Desserts"],
+      avgRating: 4.4,
+      veg: true,
+      parentId: "364",
+      avgRatingString: "4.4",
+      totalRatingsString: "1K+",
+      sla: {
+        deliveryTime: 29,
+        lastMileTravel: 1.8,
+        serviceability: "SERVICEABLE",
+        slaString: "29 mins",
+        lastMileTravelString: "1.8 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-18 00:00:00",
+        opened: true,
+      },
+      badges: {
+        imageBadges: [
+          {
+            imageId: "v1695133679/badges/Pure_Veg111.png",
+            description: "pureveg",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "pureveg",
+                  imageId: "v1695133679/badges/Pure_Veg111.png",
+                },
+              },
+            ],
+          },
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/falahaar-the-pavillion-mall-shivajinagar-pune-35370",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "200980",
+      name: "Irani Cafe",
+      cloudinaryImageId: "lwysdaf2bdkz40mqtjbr",
+      locality: "Prabhat Road",
+      areaName: "Erandwane",
+      costForTwo: "₹200 for two",
+      cuisines: ["Bakery", "Snacks", "Fast Food", "Desserts", "Beverages", "Indian", "Street Food"],
+      avgRating: 4.5,
+      parentId: "4057",
+      avgRatingString: "4.5",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 35,
+        lastMileTravel: 2.7,
+        serviceability: "SERVICEABLE",
+        slaString: "35 mins",
+        lastMileTravelString: "2.7 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:45:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/irani-cafe-prabhat-road-erandwane-pune-200980",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "5624",
+      name: "Subway",
+      cloudinaryImageId: "63178e3e64d503a479f2a2048a474552",
+      locality: "FC Road",
+      areaName: "Deccan Gymkhana",
+      costForTwo: "₹350 for two",
+      cuisines: ["Healthy Food", "Salads", "Snacks", "Desserts", "Beverages"],
+      avgRating: 4.3,
+      parentId: "2",
+      avgRatingString: "4.3",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 46,
+        lastMileTravel: 2.2,
+        serviceability: "SERVICEABLE",
+        slaString: "46 mins",
+        lastMileTravelString: "2.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:59:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/subway-fc-road-deccan-gymkhana-pune-5624",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "263939",
+      name: "The Brooklyn Creamery - Healthy Ice Cream",
+      cloudinaryImageId: "b1b35780a9b1dfeb26d680506d494eaa",
+      locality: "Ghole Road",
+      areaName: "Shivaji Nagar",
+      costForTwo: "₹200 for two",
+      cuisines: ["Desserts", "Ice Cream"],
+      avgRating: 4.6,
+      veg: true,
+      parentId: "236673",
+      avgRatingString: "4.6",
+      totalRatingsString: "500+",
+      sla: {
+        deliveryTime: 32,
+        lastMileTravel: 2.2,
+        serviceability: "SERVICEABLE",
+        slaString: "32 mins",
+        lastMileTravelString: "2.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:00:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "brand",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "brand",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/the-brooklyn-creamery-healthy-ice-cream-ghole-road-shivaji-nagar-pune-263939",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "18808",
+      name: "Faasos - Wraps, Rolls & Shawarma",
+      cloudinaryImageId: "af33b81798b11deba338e94b7585d348",
+      locality: "Shaniwar Peth FC",
+      areaName: "Shaniwar Peth",
+      costForTwo: "₹200 for two",
+      cuisines: ["Kebabs", "Fast Food", "Snacks", "American", "Healthy Food", "Desserts", "Beverages"],
+      avgRating: 4.3,
+      parentId: "21809",
+      avgRatingString: "4.3",
+      totalRatingsString: "10K+",
+      sla: {
+        deliveryTime: 26,
+        lastMileTravel: 0.4,
+        serviceability: "SERVICEABLE",
+        slaString: "26 mins",
+        lastMileTravelString: "0.4 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:59:00",
+        opened: true,
+      },
+      badges: {
+        textExtendedBadges: [
+          {
+            iconId: "guiltfree/GF_Logo_android_3x",
+            shortDescription: "options available",
+            fontColor: "#7E808C",
+          },
+        ],
+      },
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {
+            badgeObject: [
+              {
+                attributes: {
+                  description: "",
+                  fontColor: "#7E808C",
+                  iconId: "guiltfree/GF_Logo_android_3x",
+                  shortDescription: "options available",
+                },
+              },
+            ],
+          },
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/faasos-wraps-rolls-and-shawarma-fc-shaniwar-peth-pune-18808",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "395169",
+      name: "McDonald's Gourmet Burger Collection",
+      cloudinaryImageId: "oh5peivnub6ziiy9ia5o",
+      locality: "JM Road",
+      areaName: "Shivajinagar",
+      costForTwo: "₹600 for two",
+      cuisines: ["Burgers", "Beverages", "Cafe", "Desserts"],
+      avgRating: 4.3,
+      parentId: "10761",
+      avgRatingString: "4.3",
+      totalRatingsString: "100+",
+      sla: {
+        deliveryTime: 37,
+        lastMileTravel: 2.1,
+        serviceability: "SERVICEABLE",
+        slaString: "37 mins",
+        lastMileTravelString: "2.1 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-18 01:45:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/mcdonalds-gourmet-burger-collection-jm-road-shivajinagar-pune-395169",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "25628",
+      name: "Barista Coffee",
+      cloudinaryImageId: "9374825cb68f0263bc11702df311f99d",
+      locality: "Law College Road",
+      areaName: "Erandwane",
+      costForTwo: "₹350 for two",
+      cuisines: ["Beverages", "Snacks", "Desserts"],
+      avgRating: 4.4,
+      parentId: "2359",
+      avgRatingString: "4.4",
+      totalRatingsString: "500+",
+      sla: {
+        deliveryTime: 29,
+        lastMileTravel: 3.6,
+        serviceability: "SERVICEABLE",
+        slaString: "29 mins",
+        lastMileTravelString: "3.6 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/barista-coffee-law-college-road-erandwane-pune-25628",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "428100",
+      name: "Havmor Havfunn Ice Cream",
+      cloudinaryImageId: "49845b494b1e637209b7bd1c7ca184ad",
+      locality: "LB Bhopatkar Path",
+      areaName: "Swargate",
+      costForTwo: "₹400 for two",
+      cuisines: ["Ice Cream", "Ice Cream Cakes", "Desserts", "Beverages"],
+      avgRating: 4.6,
+      veg: true,
+      parentId: "94570",
+      avgRatingString: "4.6",
+      totalRatingsString: "50+",
+      sla: {
+        deliveryTime: 28,
+        lastMileTravel: 1.2,
+        serviceability: "SERVICEABLE",
+        slaString: "28 mins",
+        lastMileTravelString: "1.2 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:18:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/havmor-havfunn-ice-cream-lb-bhopatkar-path-swargate-pune-428100",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "25396",
+      name: "Oven Story Pizza - Standout Toppings",
+      cloudinaryImageId: "b39e4312ae2e9f65c3b58faaed83f488",
+      locality: "Shaniwar Peth",
+      areaName: "Narayan Peth",
+      costForTwo: "₹400 for two",
+      cuisines: ["Pizzas", "Pastas", "Italian", "Desserts", "Beverages"],
+      avgRating: 4.6,
+      parentId: "3534",
+      avgRatingString: "4.6",
+      totalRatingsString: "5K+",
+      sla: {
+        deliveryTime: 26,
+        lastMileTravel: 0.4,
+        serviceability: "SERVICEABLE",
+        slaString: "23-29 mins",
+        lastMileTravelString: "0.4 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 23:59:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/oven-story-pizza-standout-toppings-shaniwar-peth-narayan-peth-pune-25396",
+      type: "WEBLINK",
+    },
+  },
+  {
+    info: {
+      id: "412737",
+      name: "Sweet Bengal",
+      cloudinaryImageId: "i39qgflu6azefcqs5phc",
+      locality: "Dhole Patil Road",
+      areaName: "Dhole Patil Road",
+      costForTwo: "₹150 for two",
+      cuisines: ["Sweets", "Indian", "Snacks", "Desserts"],
+      avgRating: 4.5,
+      veg: true,
+      parentId: "1250",
+      avgRatingString: "4.5",
+      totalRatingsString: "1K+",
+      sla: {
+        deliveryTime: 36,
+        lastMileTravel: 4.1,
+        serviceability: "SERVICEABLE",
+        slaString: "36 mins",
+        lastMileTravelString: "4.1 km",
+        iconType: "ICON_TYPE_EMPTY",
+      },
+      availability: {
+        nextCloseTime: "2024-01-17 22:00:00",
+        opened: true,
+      },
+      badges: {},
+      isOpen: true,
+      type: "F",
+      badgesV2: {
+        entityBadges: {
+          imageBased: {},
+          textBased: {},
+          textExtendedBadges: {},
+        },
+      },
+      aggregatedDiscountInfoV3: {
+        header: "₹120 OFF",
+        subHeader: "ABOVE ₹199",
+        discountTag: "FLAT DEAL",
+      },
+      differentiatedUi: {
+        displayType: "ADS_UI_DISPLAY_TYPE_ENUM_DEFAULT",
+        differentiatedUiMediaDetails: {
+          mediaType: "ADS_MEDIA_ENUM_IMAGE",
+          lottie: {},
+          video: {},
+        },
+      },
+      reviewsSummary: {},
+      displayType: "RESTAURANT_DISPLAY_TYPE_DEFAULT",
+      restaurantOfferPresentationInfo: {},
+    },
+    analytics: {},
+    cta: {
+      link: "https://www.swiggy.com/restaurants/sweet-bengal-dhole-patil-road-pune-412737",
+      type: "WEBLINK",
+    },
+  },
+];
+export default dataList;
